Adding tests

+++ modules/robin.py
@@ -5,7 +5,7 @@ from discord.utils import get
 
 
 class Robin(commands.Cog):
-    def __init__(self, bot, conn=None):
+    def __init__(self, bot=None, conn=None):
         self.bot = bot
         self.conn = conn
         logger.info(f"Class {type(self).__name__} initialized from Robin class")
@@ -37,7 +37,7 @@ class Robin(commands.Cog):
             msg_out = msg_out.replace(nogo, '_')
         return(msg_out)
 
-    async def _feedback(self, ctx, msg: str, delete_after=None, delete_message=False) -> str:
+    async def _feedback(self, ctx=None, msg='', delete_after=None, delete_message=False) -> str:
         """
         Send feedback to the user after a message is posted.
         The original message can be deleted.
@@ -48,13 +48,14 @@ class Robin(commands.Cog):
           delete_after:   how long to wait to delete the feedback message (default keep)
           delete_message: delete the original message (default keep)
         """
-        await ctx.send(content=msg, delete_after=delete_after)
-        if delete_message:
-            try:
-                await ctx.message.delete()
-            except Exception as e:
-                logger.info(f"message deletion failed {e}")
-                return f"message deletion failed {e}"
+        if ctx is not None:
+            await ctx.send(content=msg, delete_after=delete_after)
+            if delete_message:
+                try:
+                    await ctx.message.delete()
+                except Exception as e:
+                    logger.info(f"message deletion failed {e}")
+                    return f"message deletion failed {e}"
         return "feedback sent successful"
 
     def _getusermap(self, id):
@@ -71,7 +72,6 @@ class Robin(commands.Cog):
         cur.execute(query, [id])
         row = cur.fetchone()
         usermap = {'Id': row[0], 'discordid': row[1], 'discordalias': row[2], 'gsheetalias': row[3]}
-        logger.info(f"usermap: discordid->{usermap['discordid']}, discordalias->{usermap['discordalias']}, gsheetalias->{usermap['gsheetalias']}")
 
         return(usermap)
